Add lease notifications to OH::Storage interface.
Add methods to OH::Storage to indicate that a bucket should be leased
and unleased. These don't necessarily immediately lease or lock the
bucket, but rather indicate that storage requests may occur and
leasing may be necessary. The releaseBucket is particularly useful
since leases may be acquired on demand, but should be released as soon
as possible.

Add stub implementations in the OH::Storage implementation
classes. Update HostedObject and unit tests to use the new interface.