Add lease support to SQLite storage implementation.
Add leases to SQLite storage implementation. This is relatively
straightforward since SQLite has transactions. The implementation uses
a single lease key-value pair per bucket which indicates the current
owner and expiration of the lease. We use SQLite transactions to
acquire and update the lease safely. Leases are first acquired when
the first request is made for a bucket and released when
releaseBucket() is called. A single renewal queue and timer are used
to schedule the next processing of renewals. An option controls the
standard lease duration so a tradeoff can be made between renewal
overhead and recovery time after a failure that causes a lease to not
be cleaned up.