Fix check for self in proximity removal.
We check whether our own presence is being removed from the proximity
result set so we can ignore it, avoiding removing our own ProxyObject
and ensuring we always have a copy of our own presence. This fixes the
check, which was previously just checking HostedObject's map of
presences. This check was incorrect since we would ignore removal of
*any* of the HostedObject's presences rather than just the querying
presence. This fixes the proximityAdded Emerson unit test.

This also removes an unnecessary check for the ProxyManager. We look
up the ProxyManager outside the loops handling additions and removals
so there's no need to do it again inside the removal loop.