Handle releasing leases on exit properly.
Because we might call destroy on HostedObjects only when some objects
that hold pointers to them are finally being deleted, their final
cleanup might happen after the destruction of some objects important
for their operation. This handles the release of storage buckets
properly by both releasing them earlier (in stop(), when it is called
first) to ensure the buckets actually get released, and not posting a
handler once the IOService they run on becomes invalid.