Add Commander reporting of properties, settings and state of Proximity implementations.
Add a command which gets basic properties of the Proximity
implementation and a bit of information about its current state. This
is useful for getting an overview of the current state of the query
processor. This includes the type of query processor, settings like
whether it splits static/dynamic, and state like the number of objects
and queries being tracked and the outstanding result message queue
size.