Remove FindPROX.cmake and use ADD_SUBDIRECTORY() instead.
Use libprox as a direct dependency through CMake rather than treating
it like an external library we need to search for. Pass in settings to
turn off the proxsim binaries and ensure we use the maximum size
setting for RTrees. Extract the include directories and extra
definitions to make sure we setup the buld to match the configuration
of libprox.