Rework ProcSet so it can be controlled externally.
Refactor ProcSet so a user can setup processes and control the process
externally. While a blocking wait() method is still available, it is
simple enough that writing your own version of it that does other
things while during the run is possible. This should allow tests to
start up processes and interact with them during the run.