Make tests fail directly on timeout.
Previously we inserted some special tag into the program's log to
indicate a timeout and then caught it by analyzing the
output. Instead, just make these failures happen directly. This
requires introducing a needs_hup option for CSVTests because
ProximityAddedTest currently requires the timout to expire and a
SIGHUP to be sent, which it previously made not an error by removing
it from the list of errors to check for in the output.