Use liveness on posted Emerson iStop calls.
Since they are posted, it's possible for the object to be forcefully
destroyed before they are handled. This also means the destructor
needs to be careful to take care of invoking iStop if it hasn't been
hit yet to make sure the necessary cleanup occurs even if it gets
destroyed before the callback gets invoked.