Fix forced disconnection from space servers.
Some additions for session sequence numbers had resulted in a bit of
code in disconnect() which tried to get the sequence number before
checking that the connection was still alive, resulting in an
assertion since the session sequence number wasn't valid. Moving the
call to get the sequence number after the check for a valid connection
clears up this issue. Fixes #497.