Fix number of messages reported by proximity implementations.
The previous approach was using the size of the queues used to
exchange messages between threads, but that code has since been
changed for object queries and object host queries to queue results
for each stream instead of together in a single queue. This just
switches to looping through those buffers and adding up the
outstanding messages.