Cancel mesh parsing when asset loading is cancelled.
Add handles as output of calls to OgreRenderer::parseMesh and allow
those handles to cancel the parse requests. They still go through the
queue, but do no work.

Make AssetDownloadTask track the handle and call cancel() on it when
the whole task is cancelled or destroyed. This works with the existing
download planner code and now results in much less parsing work done,
and lower latency on later parse requests, when some initial models
(top-level aggregates) start of the loading process and then get
removed.