Move method for computing visuals' names into a utility.
Move the method we use to compute a visual's name in Ogre into a
utility. It uses the same basic approach of hashing the mesh file's
hash and all texture hashes, but uses the TextureBindingsMap instead
of working with the AssetDownload. It also now prepends "vis_" to the
front so it's easy to print out the ID and tell that it's a visible.