Change computation of Ogre resource names to use hashes to enable reuse.
Change computation of Ogre resources (materials, billboard materials,
and skeletons) to use hashes of their contents instead of relying on
the name/URI of the parent mesh. This guarantees uniqueness while not
depending on the parent, enabling reuse of the same materials across
different meshes. This is beneficial both because we don't generate
duplicate materials unnecessarily and because by using the same
material we can save rendering state change costs as Ogre will group
objects with the same materials together during rendering.