Fix out-of-bounds access in BulletObject::computeCollisionShape.
We were looping over the indices in a mesh in sets of three and then
accessing them but only checking that the first one was within
range. Since we got a list with only 2 indices, this caused an
out-of-bounds error. Check that all three are in range instead of just
the first.

Fixes #502.