Pass compile definitions flag through ADD_PLUGIN_TARGET.
Pass COMPILE_DEFS_OPT to ADD_PLUGIN_TARGET to be added to plugins'
TARGET_PROPERTIES. Currently this option only adds
SIRIKATA_DEBUG_BUILD to the debug target's definitions. This became an
issue because some objects (e.g. SerializationCheck) change size
depending on this configuration. When the variable-sized object was
used in a library and a plugin, the mismatch caused memory
corruption/invalid offsets.