Fix adding more objects to an asset that's already loaded.
A very recent commit added an assertion that was invalid since we were
relying on re-downloading and processing assets even if it was already
loaded. This changes the behavior by deciding between a variety of
actions depending on the current state of the Asset's loading
process. Now no re-downloading occurs, using the asset is handled
immediately if it's already loaded, and other cases just let the
existing loading process handle the additional object.