Add basic support for tree recreation in the OH query processor.
Generate notifications (object added, property updated, object
removed) in OHLocationServiceCache for aggregates since we need that
data to make it to the query handler to reconstruct the tree. Pass on
and handle both the aggregate and parent properties properly. Make the
ObjectQueryHandlers reconstruct the tree instead of just using results
as objects.

Temporarily disable removal of nodes and objects to get things into a
working state. Without this, refinements cause the parent nodes we
want to insert into to be missing. The real fix to this problem is to
change the way results are reported, either marking the removal as
part of a refinement or just not including the removal (since now the
query is for tree replication, not to generate a result set).