Make tree replication behave properly.
Uncomment the object removal code so that tree replication works
properly again, based on an updated protocol from libprox which now
generates the right set of events instead of relying on the existing
result set (cut) reporting. This seems to be behaving well except that
since there are no additions upon coarsening, removals can cause us to
never get results back because nothing triggers
re-refinement. However, this is really just a symptom of not yet
having a good approach to deciding when to request refinements.