Add start of new manual query management strategy.
Use AggregateListener interface to pickup notifications about when
nodes are observed/fully unobserved. Forward this information from the
ObjectQueryHandler managing the tree to the ServerQueryHandler
interacting with the space server. Currently, upon observation we
request refinement.

Also update libprox to a version which handles removals, and tree
replication in general, better so this can be exercised -- the basic
test is to generate enough removals to cause the cut to be lifted up
and above where rearranging is happening and require re-refinement.

Coarsening isn't included yet, but will include a timeout after the
node is no longer observed, after which we'll request a coarsening.