Add HTTP ServerIDMap implementation.
The current version is very simple, specifying server, port, and path,
and requesting the server address using a query parameter. The
implementation could be extended (support JSON responses, substitution
in the path parameter instead of a query parameter, etc), but the
basics function if you can serve up an HTTP response that varies based
on the query parameter.

Fixes #201.