Properly handle proxies created before connection event handling in JS.
With the new manual querying code, queries can get setup and start
triggering events within the stream connected callback. Even without
this, because handling of the connection event is posted across
strands, it's possible to get query events and therefore generate
proxies before the JS code has set itself up as a listener. To handle
this properly, we need to add all the already created objects when we
setup the new connection/ProxyManager.