Fix calls to featureObject ProxManager methods within presence event callbacks.
Add a hook to allow for system functions to be invoked before any user
code is invoked for a presence connection callback. Use this to change
the implementation of the featureObject proxManager to ensure it gets
its per-presence data setup before user code is invoked, ensuring any
methods on it invoked with a user callback are handled properly. Fixes