Remove SILOGNOCR.
It was misleading since we put the carriage return in there in order
to make sure we'd get the full line of output all at once to avoid
issues with multiple threads mixing up their output. Change the couple
of locations where it was used to not require splitting output into
two steps since it wasn't any improvement to be able to do so.