Clean up dead code in OgreSystemMouseHandler.
mouseOverWebView was being called for all the mouse events, and for
some reason performed a ray trace which had no effect. Just remove it
and all calls to it. Also clear out some data we used to track when
object movement was handled in the mouse handler, but which is now
unnecessary.