Add notification to LocationServiceCache of new imposters.
Update prox to get new version of LocationServiceCache which is
notified of new imposters immediately. Update our implementations to
use this information. Not having this caused the use of dynamic
objects to break manual querying as some nodes were not properly
replicated since they couldn't be found in LocationServiceCache. We
still fail to handle dynamic objects properly, but the correct events
are at least generated.