Report index/tree properties with manual query results.
Use reported handler IDs from libprox and information we already had
about the query handlers (static/dynamic and that they are from the
local server) to generate index/tree properties in results. The raw ID
is indicated with all results, but the full information is only
generated when the root of an index is initially reported.