Fix object query update methods to ignore updates from non-queriers.
We need to forward all possible updates to queries, including for
implicit parameters like location and bounds, to the prox thread
because we don't know in the main thread whether the query
exists. However, this was causing us to track extra state for
non-queriers because we created the tracking state (sequence number in
a map in LibproxProximity and lists of queries against individual
replicated indexes in manual_query's ObjectQueryHandler) without
checking if creating the new query was really necessary. In both cases
we wouldn't create an unnecessary query in the libprox query handler,
but we were unnecessarily tracking extra state. Instead, we now check
whether one of the changes was an update to the explicit parameters
(solid angle, max number of results) and ignore updates for queries
that don't yet exist and which aren't updating either of those
parameters.