Throw exception on invalid options after all plugins are loaded.
Add a control parameter to the various ParseOptions functions provided
by CommonOptions to allow toggling whether unknown options are
accepted. Use this to only allow unregistered options on the first
round of parsing, when we're extracting basic information like
plugins. In the second round, we should have loaded all plugins and
registered their options, so we can safely throw errors when an
unknown option is encountered. Fixes #513.