Use Poller to clean out data for truly orphaned LocUpdates.
In manual_query's ServerQueryHandler, we have to create an
OrphanLocUpdateManager to handle orphaned LocUpdates. We also create a
LocationServiceCache at the same time. If a prox update didn't ever
arrive, we'd never process that cache and clear it out.

To handle this, we just keep track of a list of
OrphanLocUpdateManagers that we created for this reason and
periodically process them, destroying them if they have no valid
orphaned updates left due to timeouts but they still haven't gotten a
prox update that caused them to have contents in their location cache.