Make PintoServerQuerier interface generic.
Modify PintoServerQuerier interface so it only assumes it will get
reports from each server on the region covered by the server, the
largest object on the server, and generic query updates. The query
updates are just byte strings instead of being specific to solid angle
queries. They could encode any kind of update -- e.g. replacing a
solid angle query entirely or just issuing commands for cut
management.

Correspondingly, make the PintoServerQuerierListener interface more
generic and complete by using ProximityUpdates to report results
instead of just two callbacks for add/remove with ServerIDs. This lets
us do more than just return a list of results. This is useful for
being able to replicate the top-level tree, but also if we wanted to
make top-level pinto generate aggregates since we would then be able
to return additional properties with the top-level pinto results.

Both of these are a bit odd, but are the simplest ways to make these
interfaces generic. For the querier interface, using a string means we
need to serialize/deserialize update requests just to pass them
between two classes running on the same server because the pinto
querier object and proximity interfaces are decoupled. For the
listener, we're using proximity updates, which really only deal with
object identifiers, so we need to shoe-horn the server identifiers
into them.