Add ManualPintoManager and adjust MasterPintoManualServerQuerier to work with it.
Add ManualPintoManager which uses a manual query processor, accepting
json-encode commands to traverse the tree and returns replicated tree
nodes instead of individual servers as results. This also required a
few changes to PintoManagerLocationServiceCache to expose data that is
now needed to return those complete results.

Also adjust MasterPintoManualServerQuerier to work with this interace,
which is actually a drastic simplification since it assumes it's
client (Proximity implementation) will maintain the replicated tree
and decide how to traverse. To make this work, the interface for
MasterPintoServerQuerierBase needed to be made a bit more generic,
avoiding MasterPinto.pbj messages and using opaque byte-string
messages instead.