Refactor utility loc and pinto related classes into new libpintoloc library.
Extract a bunch of classes that are going to be useful on both the
space server and object host for performing query processing, and
tracking loc state when data is replicated, into a new libpintoloc
library. Currently its all just existing utility code, such as
LocUpdate and its implementations, OrphanLocUpdateManager refactored
to not rely on OH-specific classes, and OHLocationServiceCache renamed
to ReplicatedLocationServiceCache. The TimeSynced interface and an
implementation using the ObjectHost is added to remove the one
dependency the OrphanLocUpdateManager had on liboh.