Fix interface for injecting querier events into ManualReplicatedClient.
The existing interface only took an object ID, but since an object
could appear across multiple replicated trees, we actually need to
inject both the ProxIndexID and ObjectReference. This changes that
interface and updates related classes. ObjectQueryHandler needs to now
maintain an extra index of ProxAggregator ->
ProxIndexID. ManualReplicatedClient needs to use a combination of
ProxIndexID and ObjectReference for unobserved timeouts to make sure
it only coarsens the node in a single index. As result, this also now
includes the index in refine/coarsen requests.