Fix thread safety of RedisObjectSegmentation.
Add locking around hiredis operations in
RedisObjectSegmenetation. hiredis doesn't seem to be thread-safe.
Even if it was it could also generate read and write event callbacks
in a way we weren't expected: when we call redisAsyncHandleRead (or
Write), events callbacks of both types can be generated
(e.g. HandleRead would generate addWrite and delWrite events). Even
though there was only ever one read or write ASIO callback invoked at
a time, this could lead to conflicts on the mReading and mWriting
variables.

This patch just adds locks around all the redis commands and makes
sure the mReading and mWriting variables are lock protected. An
alternative approach would be to use a strand since only external
commands (add and remove object routing entries) would need to be
posted across to the strand, but this was a simpler solution. We can
switch to strands if this becomes a performance problem.