Make ManualPintoManager properly extract ServerIDs from refine/coarsen commands.
Previously it was checking for ints, but since clients (space servers)
see them as ObjectReferences, they'll actually come back as
Strings. Its easier to make ManualPintoManager handle all the
conversions since it allows us to easily reuse the existing
ManualReplicatedClient code. Change it to check for strings and then
parse the UUID and then convert it to a ServerID.