Start manual queries on space server with top-level pinto only.
Start manual queries on the space server at the root of the top-level
pinto tree only, refining them to other trees when the reach leaves in
the top-level tree.

This should work, but currently fails in practice because refinement
requests are being discarded as they seem to be getting processed
before more results are returned from top-level pinto, causing them to
be discarded, so refinement never actually occurs.