Fix unsafe tryRemove in ReplicatedLocationServiceCache.
At some point there was no notification posted to the other strand so
the tryRemove was appropriate, but it had made it into an unsafe
location where the iterator was used unsafely after the tryRemove
call. Just removing it is fine since we need to keep tracking the
object until the notifyObjectRemoved call completes.