Make loc updates from other servers update corresponding replicated trees in LibproxManualProximity.
This type of update requires access to the raw update since it needs a
list of indexes to apply the update too, as well as the originating
server in order to find which indexes are even candidates. To get raw
access, add a new event to LocationServiceListener which is applied
before the update is applied within LocationService.

This change makes sure that the replicated trees (and
ReplicateLocationServiceCache) which handle queries are up-to-date and
match the data in LocationService. Previously the data would just fall
out of date and query results would not be properly updated.