Add daemon and pid-file options.
Add daemon option for requesting that the process fork, detach from
the parent and stdin/out/err, and run in the background. Add a
pid-file option to request that it write it's pid. Collect these steps
and setting the log file output into a single method,
DaemonizeAndSetOutputs and call it in all binaries before our first
output, ReportVersion.