Fix thread safety when adding/removing replicas to LocationService.
In LibproxManualProximity results from other servers need to do two
things: add/remove replicas from the main LocationService and notify
the replicated trees/query handlers of the changes. These need to
happen in different threads. The LocationService operations must be in
the main strand and the replicated tree operations need to be in the
prox strand. This fixes these updates so they pass through both
strands doing the relevant work in each. Fixes a crash that was
popping up randomly under sufficient load.