Add ManualReplicatedRequestManager to track update requests that fail to apply.
Requests to update a cut may fail if the server is still waiting for
another server to return results (e.g. an OH says to refine a
replicated tree to server A, but server A is still waiting on it's
refinement command to server B to complete). To handle this, make the
server save any failed requests with a timeout and try to reapply them
when an update is received that provides new children for that node.