Properly handle aggregates missing in loc cache.
Sometimes an aggregate can be added and removed before queryHasEvents
is invoked. Previously we just filtered these objects, but that
doesn't work since if it was an aggregate then any operations with
children may get through but then reference a non-existant
node. Instead, we keep track of which objects we're ignoring and
filter any operations including them as a parent. This should be fine
because whether they are objects or aggregates, they will be either
permanently destroyed or just moved under some other node before we
exhaust the current set of events.