Cancel unobserved timeouts when objects/aggregates are removed.
The ManualReplicatedClient would sometimes request coarsening
incorrectly because of sequences of events such as observed -> remove
process triggers 0 observers -> register unobserved timeout -> later
add back -> unobserved timeout expires. This adds a method to notify
the ReplicatedClient when an object is being completely removed and
removes any outstanding timeouts.

Although the ReplicatedClient was already aware of removals since it
received prox updates, this needs to occur at a specific time: it has
to occur after the queriers working over the replicated tree have
finished removing the object. If the removal of timeouts happens
earlier, then the 0 observers event during removal will trigger the
timeout to be added back. Instead, we use the aggregateDestroyed event
and a new method in ReplicatedClient to inject the event to ensure
another timeout can't be added after we remove any outstanding ones.