Fix CBRLocationServiceCache to ensure additions/removals don't cause incorrect removal of objects.
A recent changed moved the marking of existance of objects into the
thread that invokes the addition. Sometimes doing this but keeping the
removal processing in the other strand can cause objects to be
incorrectly removed because of successive additions/removals that are
then processed in the incorrect order. Instead, move both to be
processed immediately, but use refcounts to ensure they remain alive
until the corresponding handler in the other thread actually
executes. This is just moving closer to how other LocationServiceCache
implementations are working where more processing happens immediately
and only notifications to listeners that care to be notified in a
separate strand are handled there.