Add InstanceMethodNotReentrant helper and verify queryHasEvents is not called recursively.
Add InstanceMethodNotReentrant helper which uses an instance variable
in debug mode to ensure two calls to an instance method are not on the
stack at the same time. Use this for proximity queryHasEvents calls,
which, if they call other prox methods could accidentally trigger
another event callback. Fix a problem in LibproxManualProximity that
this exposes, where registering a query against a new server when the
bottom of top-level Pinto's tree is reached causes an immediate
callback while still in the middle of processing the one event, let
alone any other events that may also need processing. The fix is to
just defer (un)registration by posting the call.