Remove LocationService::type() and fix application of replica location updates.
Remove LocationService::type() method and the TrackingType enum. This
enum was no longer valid since it incorrectly combined two things:
local vs. replica and object vs. aggregate. Previously we did not
track the type of replicas so this made sense, but with replicated
trees that's no longer the case. Since the method was not used
publicly (only by the LocationService implementations themselves), it
can just be removed.

The remaining uses in StandardLocationService and BulletPhysicsService
required replacements. In both cases all they actually cared about was
local vs. remote. In the case of LocationUpdates, because of the way
the type() method had been implemented, we were incorrectly ignoring
replica aggregate updates because they were reported as aggregates
only, and therefore it had been assumed they shouldn't be getting
remote updates. Instead, we now just check if the object is remote or
not and then allow the update.