Make number of threads in AggregateManager configurable.
Provide options for both generation and upload threads. There are
still fixed maximum number of threads for each, but they can just be
made arbitrarily large if more threads are needed. The real defaults
are now options, aggmgr.gen-threads and aggmgr.upload-threads.

This also adds a few checks within AggregateManager to handle the case
where no generation threads are provided, i.e. we want to disable all
work associated with aggregate generation. This would probably be
better accomplished without also tracking all the state and never
clearing it, but this is a workable short-term solution.