Speed up UUID creation by about 2000x.
Instead of creating a the random generator state every time we
generate a UUID, use a thread_specific_ptr storing a random generator
so it can be reused. Setting up the random number generator is
actually the expensive part, and regenerating it is probably bad
anyway. This makes the benchmark on my machine jump from about 1K
UUIDs/s to about 2M UUIDs/s.