Don't catch SIGABRT.
We should only be getting these on assertions, and if we're violating
one it is better to die immediately, potentially getting a good stack
trace in a crash dump, rather than catch the error trying to cleanup
and exit gracefully, leaving no indication of what happened. If at
some point SIGABRT is coming from something else, we can revisit
this.