Fix conflicting retries and redirects in SessionManager.
This was a long standing bug in SessionManager where we would see an
assertion that the seqno should not be 0. It looks like it was caused
only during redirects where the entire login took longer than the
retry timeout. If we got a redirect back, we'd reset the seqno to 0,
but the connectingTo would remain the same until we got the initial
underlying connection to the new space server. If the retry callback
occurred during this time, we'd end up thinking we were still trying
to connect to the original server, but the seqno would be bad,
triggering the assertion. This explains why it only happened rarely --
you had to get redirects and the retry had to happen at just the right
time because if you already had a connection to the redirected space
server it would have been used immediately. The fix is to check both
the server we're trying to connect to and that the seqno hasn't been
reset.