Fix thread safety issue in Redis OSeg.
Some of the callbacks invoked by read/write handlers from hiredis were
accessing mOSeg and the timeouts queue, but the locks they use only
protect hiredis access. Post to the main strand to actually process
these events so we don't need to lock them everywhere. The affected
events are only for writing new and migrated objects, which should be
relatively rare compared to lookups, so posting each individually
shouldn't be a performance issue.