Add migration query test.
Add test for queries after migration, ensuring they see results
returned to them again. This revealed an issue where an iterator
became invalidated during a callback, so we need to exit that loop if
when we get control back we see that the entity is being killed. This
also revealed that some JSVisibleManager callbacks could be invoked
after it had been destroyed because they were posted without liveness
checks. This also fixes that issue.