Add Emerson loadMesh option to download all referenced resources.
Add an option to Emerson's loadMesh methods on visibles and presences
that requests the entire asset (including dependent resources like
textures) be downloaded. This probably wouldn't be used normally, but
is useful for performing downloads like a client would without loading
up the Ogre plugin.

The modification to the Emerson code is limited to adding the flag and
using AssetDownloadTask instead of ResourceDownloadTask when it is
enabled. However, to use AssetDownloadTask it needed to be factored
out of the Ogre code. Move it into libmesh and remove dependencies on
OgreRenderer by passing in a TransferPoolPtr directly and adding a new
interface, Mesh::ParserService which performs async parsing of
meshes. Make OgreRenderer and JSObjectScriptManager implement this
interface, which is just a single function providing functionality
they both already had.