Initialize Berkelium earlier so forked process is smaller.
Do some initialization of WebViewManager and Berkelium as early as
possible so the forked process uses less memory. Only the small amount
of code to get Berkelium::init called is used here, the rest remains
initialized at the normal time. Doing this earlier avoids wasting a
bunch of memory if the original process continues to modify a bunch of
already allocated memory, causing copies of pages to be generated.