Add ClutterRenderer skeleton.
Add skeleton for a ClutterRenderer, a new Simulation that renders to
screen using Clutter. The skeleton includes setup and teardown
code. This has to be careful to handle threading properly because
Clutter is "thread-aware" but not thread-safe. Because of this, a
separate clutter renderer thread is used and events need to be
injected into it. The shutdown process demonstrates this, using
clutter_threads_add_idle to request that the main loop terminate.

Also add a CMake module to find Clutter, which will be used for
rendering, incorporate the plugin into the CMakeLists.txt build
script, only building when Clutter is available, and a simple Emerson
class for interacting with the renderer.