Add command handling support to objects and their scripts.
Add a new command for object hosts that dispatches the command to a
particular object for handling. Objects can't do anything with these
directly. Instead, also add a function to the ObjectScript interface
that handles these requests. Add the necessary infrastructure to the
Emerson plugin for a script to register a callback which is invoked
for each of these requests. Currently, that interface assumes a
synchronous-style interface because it greatly simplifies the
implementation because the associated command state does not have to
be saved somewhere and no callback into native code needs to be
provided or handled.