Add basic aggregate processing to TwitterAggregateManager.
When aggregates are processed, download and parse all their children's
information. Use a cache to avoid downloading more than
necessary. Then, as a temporary type of aggregate, create a subset of
the tweet data, serialize the JSON, and upload it to a server
(currently using a simple HTTP request instead of the OAuth approach
because that makes the storage implementation simpler). Finally, with
the new data in place, construct and set a new URL for the aggregate,
allowing processing of parents to proceed.