Add support for dynamic selection of per-node query processor data.
Update libprox to a new version that allows the type of per-node RTree
data to be extensible and specified through template
parameters. Adjust the QueryHandlerFactory to take an additional
parameter to select the type of per-node data, making this a dynamic
selection. Also update all allocations of query processors to work
through the factory, necessitating the addition of a version for
manual query handlers. Update all calls to the factory, adding options
to control the type of data used; default to maxsize, as the
compile-time settings already did.