Make QueryHandlerFactories extensible.
Move QueryHandlerFactory into libpintoloc, which is now a more
appropriate location. Change the implementation of the factories to
allow registration, provide defaults, and work through classes for
each type of query handler and templated on the
SimulationTraits. These classes aren't singletons because getting them
to work across multiple libraries and binaries with different export
constraints (e.g. ServerProxSimulationTraits is only used in the pinto
binary whereas ObjectProxSimulationTraits is used from multiple
libraries) is seemingly impossible. Instead, for each SimulationTraits
type, one singleton object is maintained. For
ObjectProxSimulationTraits it is maintained and exported by
libpintoloc, for ServerProxSimulationTraits it is maintained by the
pinto binary.