Add custom query data support and a Twitter term+region query handler.
Add custom query data support to the various proximity classes. First,
this requires the ability to pass custom query data to, and register
custom queries with, query handlers. It also adds callback support for
custom query data events on aggregates, allowing the data to be
propagated. This second step is critical for querying to work between
space servers or using replicated trees.

Add a new query handler implementation that queries by search term and
region. It uses the Twitter per-node data types that include a Bloom
filter. The query is parsed by the query handler from JSON and must
include a term and a region, but the region need only be 2D. It only
refines if it finds both the search term and the aggregate overlaps
the region. This is a fairly straightforward adaptation of the
RTreeCutQueryHandler or LevelQueryHandler with different criteria, and
additions to support the custom query format.

A few fixes were also necessary. CopyableLocUpdate was not copying the
query data properly, which wasn't noticeable until a query handler in
the OH tried to use the data. The ReplicatedLocationServiceCache wasn
not properly notifying downstream LocationUpdateListeners of query
data updates, so the data was not properly being passed into OH query
handlers. Const-ness was broken on a number of TermBloomFilter
methods. The TermBloomFilter was also missing a critical memset to 0
for the filter data in the constructor.