Make TwitterAggregateManager more robust to ordering issues.
Because we get object creation/deletion and aggregate
creation/update/deletion from different sources on different threads,
perfect ordering isn't guaranteed. Sometimes, for example, and
addChild(aggregate, child) can be called before the
objectCreated(child). This makes either function capable of creating
the entry for the child and makes sure both mark the appropriate
objects as dirty when the signal is received.