Fix removed child cleanup in TwitterAggregateManager.
When clearing out a child aggregate, we need to make sure we adjust
the list of dirty children in its parent. We were deleting an element
for the removed aggregate, but we need to remove an entire range since
it may be in the list multiple times in order to support marking the
tree as dirty while aggregates are being generated.